% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldat.R, R/alldat_dk.R
\docType{data}
\name{alldat}
\alias{alldat}
\alias{alldat}
\alias{alldat_dk}
\title{Simulated Responses to Knowledge Questions Without Don't Know}
\format{A data frame with 500 rows and 4000 variables:}
\usage{
alldat

alldat_dk
}
\description{
Simulate Pre- and Post-Process Responses to Knowledge Questions without Don't Know

Simulate Pre- and Post-Process Responses to Knowledge Questions with Don't Know
}
\keyword{data}
