% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{calculate_expected_values}
\alias{calculate_expected_values}
\title{Calculate expected values for goodness of fit test}
\usage{
calculate_expected_values(gamma_i, params, total_obs, model_type = "nodk")
}
\arguments{
\item{gamma_i}{item-specific gamma value}

\item{params}{estimated parameters for the item}

\item{total_obs}{total observations for the item}

\item{model_type}{"nodk" or "dk" model}
}
\value{
vector of expected values
}
\description{
Calculate expected values for goodness of fit test
}
