% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{coef.guess_fit}
\alias{coef.guess_fit}
\title{Extract coefficients from guess_fit}
\usage{
\method{coef}{guess_fit}(object, ...)
}
\arguments{
\item{object}{guess_fit object}

\item{...}{ignored}
}
\value{
parameter matrix
}
\description{
Extract coefficients from guess_fit
}
