% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_sim.R
\docType{data}
\name{dk_sim}
\alias{dk_sim}
\title{Simulated Responses to Knowledge Questions With Don't Know}
\format{A data frame with 500 rows and 4000 variables:}
\usage{
dk_sim
}
\description{
Simulate Pre- and Post-Process Responses to Knowledge Questions with Don't Know
}
\keyword{data}
