% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_sim_params.R
\docType{data}
\name{dk_sim_params}
\alias{dk_sim_params}
\title{Simulated Responses to Knowledge Questions With Don't Know}
\format{A data frame with 1000 rows and 5 variables:}
\usage{
dk_sim_params
}
\description{
Simulate Pre- and Post-Process Responses to Knowledge Questions with Don't Know
}
\keyword{data}
