% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{eqn1}
\alias{eqn1}
\title{Sum to 1 constraint (no DK)}
\usage{
eqn1(x, g1 = NA, data)
}
\arguments{
\item{x}{gg, gk, kk}

\item{g1}{guess}

\item{data}{transition matrix}
}
\description{
Constraints that some params sum to 1. Used Internally. For data without DK.
Functions for constraining lambdas to sum to 1 and to bound params between 0 and 1
}
\keyword{internal}
