% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_dk.R
\name{fit_dk}
\alias{fit_dk}
\title{fit_dk}
\usage{
fit_dk(pre_test, pst_test, g, est.param)
}
\arguments{
\item{pre_test}{data.frame carrying pre_test items}

\item{pst_test}{data.frame carrying pst_test items}

\item{g}{estimates of \eqn{\gamma}}

\item{est.param}{estimated parameters}
}
\value{
fit statistics
}
\description{
Calculate fit statistics for data with dk
}
\examples{
\dontrun{fit_dk(pre_test, pst_test, )}
}

