% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_nodk.R
\name{fit_nodk}
\alias{fit_nodk}
\title{fit_nodk}
\usage{
fit_nodk(pre_test, pst_test, g, est.param)
}
\arguments{
\item{pre_test}{data.frame carrying pre_test items}

\item{pst_test}{data.frame carrying pst_test items}

\item{g}{estimates of \eqn{\gamma}}

\item{est.param}{estimated parameters}
}
\value{
fit statistics
}
\description{
fit for data without dk
}
\examples{
\dontrun{fit_nodk(g, est.param, data)}
}

