% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{format_transition_matrix}
\alias{format_transition_matrix}
\title{Format transition matrix result with appropriate row and column names}
\usage{
format_transition_matrix(
  transition_list,
  item_names,
  include_aggregate = FALSE
)
}
\arguments{
\item{transition_list}{list of transition vectors}

\item{item_names}{item names in output row order}

\item{include_aggregate}{whether to add an aggregate row}
}
\value{
formatted matrix
}
\description{
Format transition matrix result with appropriate row and column names
}
