% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\docType{package}
\name{guess-package}
\alias{guess-package}
\alias{guess}
\title{\pkg{guess} adjust estimates of learning for guessing related bias.}
\description{
It implements the method discussed in \url{https://gsood.com/research/papers/guess.pdf}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/finite-sample/guess}
  \item \url{https://finite-sample.github.io/guess/}
  \item Report bugs at \url{https://github.com/finite-sample/guess/issues}
}

}
\author{
\strong{Maintainer}: Gaurav Sood \email{gsood07@gmail.com}

Authors:
\itemize{
  \item Gaurav Sood \email{gsood07@gmail.com}
  \item Ken Cor \email{mcor@ualberta.ca}
}

}
\keyword{internal}
