% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\docType{package}
\name{guess}
\alias{guess}
\alias{guess-package}
\title{\pkg{guess} adjusts estimates of learning for guessing related bias.}
\description{
It implements the method discussed in \url{http://gsood.com/research/papers/guess.pdf}
}

