% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{guessdk_lik}
\alias{guessdk_lik}
\title{guessdk_lik}
\usage{
guessdk_lik(x, g1 = x[8], data)
}
\arguments{
\item{x}{gg, gk, gd, kg, kk, kd, dd}

\item{g1}{guess}

\item{data}{transition matrix}
}
\description{
Likelihood function for data with Don't Know. Used Internally.
}
\keyword{internal}
