% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guesstimate.R
\name{guesstimate}
\alias{guesstimate}
\title{guesstimate
Calculate item level and aggregate estimates}
\usage{
guesstimate(transmat = NULL)
}
\arguments{
\item{transmat}{data.frame returned from multi_transmat}
}
\value{
estimates
}
\description{
guesstimate
Calculate item level and aggregate estimates
}

