% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-criticism.R
\name{individual_log_likelihood}
\alias{individual_log_likelihood}
\title{Calculate per-individual log-likelihood}
\usage{
individual_log_likelihood(lca_result, pre_test, pst_test)
}
\arguments{
\item{lca_result}{output from lca_cor() or lca_fit()}

\item{pre_test}{data.frame of pre-test responses}

\item{pst_test}{data.frame of post-test responses}
}
\value{
numeric vector of log-likelihoods (length = n individuals)
}
\description{
Calculate per-individual log-likelihood
}
\keyword{internal}
