% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{interleave}
\alias{interleave}
\title{Interleave vectors}
\usage{
interleave(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\description{
Interleaves two vectors. Used internally.
}
\keyword{internal}
