% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-criticism.R
\name{lca_fit}
\alias{lca_fit}
\title{Fit LCA model from individual-level data}
\usage{
lca_fit(pre_test, pst_test, ...)
}
\arguments{
\item{pre_test}{data.frame of pre-test responses}

\item{pst_test}{data.frame of post-test responses}

\item{...}{passed to lca_cor()}
}
\value{
output from lca_cor()
}
\description{
Convenience wrapper: creates transition matrix and fits model.
}
