% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-criticism.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate log-likelihood for transition data}
\usage{
log_likelihood(params, data)
}
\arguments{
\item{params}{numeric vector of length 4 (nodk) or 8 (dk)}

\item{data}{numeric vector of transition counts}
}
\value{
scalar log-likelihood
}
\description{
Calculate log-likelihood for transition data
}
\examples{
params <- c(0.4, 0.3, 0.3, 0.25)
data <- c(x00 = 10, x01 = 5, x10 = 3, x11 = 12)
log_likelihood(params, data)
}
