% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_lca.R
\name{make_person_lca_start}
\alias{make_person_lca_start}
\title{Make a joint person-level LCA starting point}
\usage{
make_person_lca_start(pre_test, post_test)
}
\arguments{
\item{pre_test}{normalized binary pre-test responses}

\item{post_test}{normalized binary post-test responses}
}
\value{
named list of class priors and item guessing probabilities
}
\description{
Make a joint person-level LCA starting point
}
\keyword{internal}
