% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{multinomial_nll}
\alias{multinomial_nll}
\title{Multinomial negative log-likelihood}
\usage{
multinomial_nll(data, probs)
}
\arguments{
\item{data}{observed cell counts}

\item{probs}{model-implied cell probabilities}
}
\value{
scalar negative log-likelihood
}
\description{
Multinomial negative log-likelihood
}
\keyword{internal}
