% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{normalize_missing_action}
\alias{normalize_missing_action}
\title{Normalize structural missingness handling}
\usage{
normalize_missing_action(missing_action)
}
\arguments{
\item{missing_action}{how structural missingness should be handled}
}
\value{
one of "omit" or "error"
}
\description{
Normalize structural missingness handling
}
\keyword{internal}
