% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{normalize_na_as}
\alias{normalize_na_as}
\title{Normalize how NA responses are classified}
\usage{
normalize_na_as(na_as)
}
\arguments{
\item{na_as}{whether NA represents an observed don't know response or
structural missingness}
}
\value{
one of "dk" or "missing"
}
\description{
Normalize how NA responses are classified
}
\keyword{internal}
