% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{normalize_responses}
\alias{normalize_responses}
\title{Normalize raw item responses}
\usage{
normalize_responses(
  x,
  na_as = c("dk", "missing"),
  missing_action = c("omit", "error")
)
}
\arguments{
\item{x}{response vector}

\item{na_as}{classification of NA responses}

\item{missing_action}{structural missingness handling}
}
\value{
character vector containing "0", "1", "d", or NA
}
\description{
Normalize raw item responses
}
\keyword{internal}
