% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_guess.R
\name{p_guess}
\alias{p_guess}
\title{Person Level Adjustment}
\usage{
p_guess(pre = NULL, pst = NULL)
}
\arguments{
\item{pre}{pre data frame}

\item{pst}{pst data frame}
}
\value{
adjusted responses
}
\description{
Adjusts observed 1s based on item level parameters of the LCA model. Currently only takes data with Don't Know.
If NAs are observed in the data, they are treating as acknowledgments of ignorance.
}
\examples{
pre_test_var <- data.frame(pre=c(1,0,0,1,"d","d",0,1,NA))
pst_test_var <- data.frame(pst=c(1,NA,1,"d",1,0,1,1,"d"))
p_guess(pre_test_var, pst_test_var)
}

