% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_lca.R
\name{person_item_response_probs}
\alias{person_item_response_probs}
\title{Person/item response probabilities}
\usage{
person_item_response_probs(gamma)
}
\arguments{
\item{gamma}{item guessing probability}
}
\value{
matrix with response pairs in rows and latent classes in columns
}
\description{
Person/item response probabilities
}
\keyword{internal}
