% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.guess_cv}
\alias{print.guess_cv}
\title{Print method for guess_cv}
\usage{
\method{print}{guess_cv}(x, ...)
}
\arguments{
\item{x}{guess_cv object}

\item{...}{ignored}
}
\value{
invisible(x)
}
\description{
Print method for guess_cv
}
