% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.guess_fit}
\alias{print.guess_fit}
\title{Print method for guess_fit}
\usage{
\method{print}{guess_fit}(x, ...)
}
\arguments{
\item{x}{guess_fit object}

\item{...}{ignored}
}
\value{
invisible(x)
}
\description{
Print method for guess_fit
}
