% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.guess_gof}
\alias{print.guess_gof}
\title{Print method for item-level goodness-of-fit diagnostics}
\usage{
\method{print}{guess_gof}(x, ...)
}
\arguments{
\item{x}{A `guess_gof` object.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print method for item-level goodness-of-fit diagnostics
}
