% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stnd_adj.R
\name{stnd_adj}
\alias{stnd_adj}
\title{Person Level Standard Guessing Correction Adjustment}
\usage{
stnd_adj(pre = NULL, pst = NULL, lucky = NULL, dk = 0.03)
}
\arguments{
\item{pre}{pre data frame}

\item{pst}{pst data frame}

\item{lucky}{change of getting the right answer by chance}

\item{dk}{Numeric. Between 0 and 1. Hidden knowledge behind don't know responses. Default is .03.}
}
\value{
list of pre and post adjusted responses
}
\description{
Adjusts observed 1s based on the standard guessing correction. 
If NAs are observed in the data, they are treating as acknowledgments of ignorance.
}
\examples{
pre_test_var <- data.frame(pre=c(1,0,0,1,"d","d",0,1,NA))
pst_test_var <- data.frame(pst=c(1,NA,1,"d",1,0,1,1,"d"))
lucky <- c(.25)
stnd_adj(pre_test_var, pst_test_var, lucky)
}

