% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stndcor.R
\name{stndcor}
\alias{stndcor}
\title{Standard Guessing Correction for Learning}
\usage{
stndcor(pre_test = NULL, pst_test = NULL, lucky = NULL)
}
\arguments{
\item{pre_test}{data.frame carrying responses to pre-test questions.}

\item{pst_test}{data.frame carrying responses to post-test questions.}

\item{lucky}{}
}
\value{
a numeric vector containing adjusted estimates of learning for each item
}
\description{
Estimate of learning adjusted with standard correction for guessing. Correction is based on number of options per question.
Requirements: separate pre-test and post-test dataframes. Why do we need dataframes? To accomodate multiple items.
The items can carry NA (missing). Items must be in the same order in each dataframe. For now we assume that respondents are posed same questions twice.
The function also takes a lucky vector - chance of getting a correct answer if guessing randomly. Each entry is 1/(no. of options).
}
\examples{
pre_test <- data.frame(item1=c(1,0,0,1,0), item2=c(1,NA,0,1,0)); pst_test <-  pre_test + cbind(c(0,1,1,0,0), c(0,1,0,0,1))
lucky <- rep(.25, 2); stndcor(pre_test, pst_test, lucky)
}

