% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{summary.guess_cv}
\alias{summary.guess_cv}
\title{Summary method for guess_cv}
\usage{
\method{summary}{guess_cv}(object, ...)
}
\arguments{
\item{object}{guess_cv object}

\item{...}{ignored}
}
\value{
invisible summary
}
\description{
Summary method for guess_cv
}
