% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{summary.guess_fit}
\alias{summary.guess_fit}
\title{Summary method for guess_fit}
\usage{
\method{summary}{guess_fit}(object, ...)
}
\arguments{
\item{object}{guess_fit object}

\item{...}{ignored}
}
\value{
invisible summary object
}
\description{
Summary method for guess_fit
}
