% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{tabulate_transition_pairs}
\alias{tabulate_transition_pairs}
\title{Tabulate transitions between pre-test and post-test responses}
\usage{
tabulate_transition_pairs(pre_responses, post_responses, include_dk = FALSE)
}
\arguments{
\item{pre_responses}{character vector of pre-test responses}

\item{post_responses}{character vector of post-test responses}

\item{include_dk}{whether to include the five don't-know transition cells}
}
\value{
named vector of transition counts
}
\description{
Tabulate transitions between pre-test and post-test responses
}
\keyword{internal}
