% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_compatible_dataframes}
\alias{validate_compatible_dataframes}
\title{Validate that two data frames have compatible dimensions}
\usage{
validate_compatible_dataframes(pre_test, pst_test)
}
\arguments{
\item{pre_test}{pre-test data frame}

\item{pst_test}{post-test data frame}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate that two data frames have compatible dimensions
}
