% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_dataframe}
\alias{validate_dataframe}
\title{Validate that input is a data frame}
\usage{
validate_dataframe(x, arg_name)
}
\arguments{
\item{x}{input to validate}

\item{arg_name}{name of the argument for error messages}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate that input is a data frame
}
\keyword{internal}
