% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_dk}
\alias{validate_dk}
\title{Validate dk parameter (knowledge behind don't know responses)}
\usage{
validate_dk(dk)
}
\arguments{
\item{dk}{numeric value between 0 and 1}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate dk parameter (knowledge behind don't know responses)
}
\keyword{internal}
