% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_equal_length}
\alias{validate_equal_length}
\title{Validate that vectors have equal length}
\usage{
validate_equal_length(vec1, vec2, name1 = "vector1", name2 = "vector2")
}
\arguments{
\item{vec1}{first vector}

\item{vec2}{second vector}

\item{name1}{name of first vector for error messages}

\item{name2}{name of second vector for error messages}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate that vectors have equal length
}
\keyword{internal}
