% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_gamma}
\alias{validate_gamma}
\title{Validate gamma parameter}
\usage{
validate_gamma(gamma)
}
\arguments{
\item{gamma}{probability parameter}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate gamma parameter
}
