% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_guessing_probability}
\alias{validate_guessing_probability}
\title{Validate guessing probabilities for standard correction}
\usage{
validate_guessing_probability(guessing_probability, n_items)
}
\arguments{
\item{guessing_probability}{vector of guessing probabilities}

\item{n_items}{number of items to validate against}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate guessing probabilities for standard correction
}
