% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_item_lca_fit}
\alias{validate_item_lca_fit}
\title{Validate an item-level latent-class fit}
\usage{
validate_item_lca_fit(fit)
}
\arguments{
\item{fit}{fitted item-level latent-class model}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate an item-level latent-class fit
}
\keyword{internal}
