% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_lucky_vector}
\alias{validate_lucky_vector}
\title{Validate lucky vector for standard correction}
\usage{
validate_lucky_vector(lucky, n_items)
}
\arguments{
\item{lucky}{vector of guessing probabilities}

\item{n_items}{number of items to validate against}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate lucky vector for standard correction
}
