% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_matrix}
\alias{validate_matrix}
\title{Validate matrix input}
\usage{
validate_matrix(x, arg_name, valid_ncols = NULL)
}
\arguments{
\item{x}{input to validate}

\item{arg_name}{name of the argument for error messages}

\item{valid_ncols}{optional vector of valid column counts}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate matrix input
}
\keyword{internal}
