% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_paired_item_names}
\alias{validate_paired_item_names}
\title{Validate names of paired item data}
\usage{
validate_paired_item_names(pre_test, post_test)
}
\arguments{
\item{pre_test}{pre-test item data frame}

\item{post_test}{post-test item data frame}
}
\value{
item names in pre-test column order
}
\description{
Validate names of paired item data
}
\keyword{internal}
