% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_person_lca_fit}
\alias{validate_person_lca_fit}
\title{Validate a person-level latent-class fit}
\usage{
validate_person_lca_fit(fit)
}
\arguments{
\item{fit}{fitted person-level latent-class model}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate a person-level latent-class fit
}
\keyword{internal}
