% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_priors}
\alias{validate_priors}
\title{Validate prior parameters}
\usage{
validate_priors(priors, expected_length, param_name)
}
\arguments{
\item{priors}{vector of prior parameters}

\item{expected_length}{expected length of priors vector}

\item{param_name}{name of parameter for error messages}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate prior parameters
}
