% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_subgroup}
\alias{validate_subgroup}
\title{Validate subgroup parameter}
\usage{
validate_subgroup(subgroup, expected_length)
}
\arguments{
\item{subgroup}{logical vector for subsetting}

\item{expected_length}{expected length to match}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate subgroup parameter
}
\keyword{internal}
