% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_transition_values}
\alias{validate_transition_values}
\title{Validate transition matrix values}
\usage{
validate_transition_values(pre_test_var, pst_test_var)
}
\arguments{
\item{pre_test_var}{pre-test variable vector}

\item{pst_test_var}{post-test variable vector}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate transition matrix values
}
