% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{with_preserved_seed}
\alias{with_preserved_seed}
\title{Evaluate an expression with a temporary random seed}
\usage{
with_preserved_seed(seed, expr)
}
\arguments{
\item{seed}{integer seed or NULL}

\item{expr}{expression to evaluate}
}
\value{
value of expr
}
\description{
Evaluate an expression with a temporary random seed
}
\keyword{internal}
