% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{zero1}
\alias{zero1}
\title{Constrain vector to [0,1] range}
\usage{
zero1(x)
}
\arguments{
\item{x}{numeric vector to constrain}
}
\value{
numeric vector with values constrained to [0,1]
}
\description{
Constrains values in a vector to be between 0 and 1. Used internally.
}
\keyword{internal}
