% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldat.R
\docType{data}
\name{alldat}
\alias{alldat}
\title{Simulated Responses to Knowledge Questions Without Don't Know}
\format{A data frame with 500 rows and 4000 variables:}
\usage{
alldat
}
\description{
Simulate Pre- and Post-Process Responses to Knowledge Questions without Don't Know
}
\keyword{data}

