% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{count_transitions}
\alias{count_transitions}
\title{Count transitions between pre and post test responses}
\usage{
count_transitions(pre_responses, pst_responses)
}
\arguments{
\item{pre_responses}{character vector of pre-test responses}

\item{pst_responses}{character vector of post-test responses}
}
\value{
named vector of transition counts
}
\description{
Count transitions between pre and post test responses
}
\keyword{internal}
