% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{eq1dk}
\alias{eq1dk}
\title{Constraints: Sum to 1}
\usage{
eq1dk(x, g1 = NA, data)
}
\arguments{
\item{x}{gg, gk, gd, kg, kk, kd, dd}

\item{g1}{guess}

\item{data}{transition matrix}
}
\description{
Constraints that some params sum to 1. Used Internally. For data with DK.
Functions for constraining lambdas to sum to 1 and to bound params between 0 and 1
}
\keyword{internal}
