% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqn1.R
\name{eqn1}
\alias{eqn1}
\title{Constraints: Sum to 1}
\usage{
eqn1(x, g1 = NA, data)
}
\arguments{
\item{x}{lgg, lgk, lkk}

\item{g1}{guess}

\item{data}{transition matrix}
}
\description{
Constraints that some params sum to 1. Used Internally. For data with DK.
Functions for constraining lambdas to sum to 1 and to bound params between 0 and 1
}
\keyword{internal}

