% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-criticism.R
\name{extract_params}
\alias{extract_params}
\title{Extract parameter matrix from lca_result}
\usage{
extract_params(lca_result)
}
\arguments{
\item{lca_result}{output from lca_cor/item_lca_fit or numeric vector}
}
\value{
matrix of parameters (rows = params, cols = items)
}
\description{
Extract parameter matrix from lca_result
}
\keyword{internal}
