% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_dk.R
\name{fit_dk}
\alias{fit_dk}
\title{fit_dk}
\usage{
fit_dk(g, est.param, data, nchisims = NULL)
}
\arguments{
\item{g}{}

\item{est.param}{}

\item{data}{}

\item{nchisims}{no. of chisq sims}
}
\value{
interleaved vector
}
\description{
Calculate fit statistics for data with dk
}
\examples{
\dontrun{fit_dk()}
}

