% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_nodk.R
\name{fit_nodk}
\alias{fit_nodk}
\title{fit_nodk}
\usage{
fit_nodk(g, est.param, data)
}
\arguments{
\item{g}{}

\item{est.param}{estimated parameters}

\item{data}{}
}
\value{
interleaved vector
}
\description{
fit for data without dk
}
\examples{
\dontrun{fit_nodk()}
}

