% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nodk.R
\name{fit_nodk}
\alias{fit_nodk}
\title{Goodness of fit statistics for data without don't know}
\usage{
fit_nodk(pre_test, pst_test, g, est.param)
}
\arguments{
\item{pre_test}{data.frame carrying pre_test items}

\item{pst_test}{data.frame carrying pst_test items}

\item{g}{estimates of \eqn{\gamma} produced from \code{\link{lca_cor}}}

\item{est.param}{estimated parameters produced from \code{\link{lca_cor}}}
}
\value{
matrix with two rows: top row carrying chi-square value, and bottom row probability of observing that value
}
\description{
For data without Don't Know, chi-square goodness of fit between true and model based multivariate distribution
}
\details{
fit_nodk
}
\examples{
\dontrun{fit_nodk(pre_test, pst_test, g, est.param)}
}
