% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_lik.R
\name{guess_lik}
\alias{guess_lik}
\title{guess_lik}
\usage{
guess_lik(x, g1 = x[4], data)
}
\arguments{
\item{x}{lgg, lgk, lkk}

\item{g1}{guess}

\item{data}{transition matrix}
}
\description{
Likelihood function for data without Don't Know. Used Internally.
}
\keyword{internal}

