% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guess_stnderr.R
\name{guess_stnderr}
\alias{guess_stnderr}
\title{guess_stnderr: standard error}
\usage{
guess_stnderr(pre_test, pst_test, nsamps = 100, seed = 31415)
}
\arguments{
\item{pre_test}{data.frame carrying pre_test items}

\item{pst_test}{data.frame carrying pst_test items}

\item{nsamps}{number of resample}

\item{seed}{random seed}
}
\value{
matrix with rows = total number of items + 1 (last row contains aggregate distribution across items)
number of columns = 4 when no don't know, and 9 when there is a don't know option
}
\description{
Bootstrapped standard errors of effect size estimates
}
\examples{
pre_test <- data.frame(pre_item1=c(1,0,0,1,0), pre_item2=c(1,NA,0,1,0))
pst_test <- data.frame(pst_item1=pre_test[,1] + c(0,1,1,0,0),
						  pst_item2 = pre_test[,2] + c(0,1,0,0,1))
\dontrun{guess_stnderr(pre_test, pst_test, 100, nsamps=10, seed = 31415)}
}

