% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guessdk_lik.R
\name{guessdk_lik}
\alias{guessdk_lik}
\title{guessdk_lik}
\usage{
guessdk_lik(x, g1 = x[8], data)
}
\arguments{
\item{x}{lgg, lgk, lgc, lkk, lcg, lck, and lcc}

\item{g1}{guess}

\item{data}{transition matrix}
}
\description{
Likelihood function for data with Don't Know. Used Internally.
}
\keyword{internal}

