% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guesstimate.R
\name{guesstimate}
\alias{guesstimate}
\title{guesstimate
Calculate item level and aggregate estimates}
\usage{
guesstimate(pre_test = NULL, pst_test = NULL, subgroup = NULL)
}
\arguments{
\item{pre_test}{data.frame carrying pre_test items}

\item{pst_test}{data.frame carrying pst_test items}

\item{subgroup}{a dummy vector identifying the subset}
}
\value{
estimates
}
\description{
guesstimate
Calculate item level and aggregate estimates
}

