% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesstimate.R
\name{guesstimate}
\alias{guesstimate}
\title{Calculate item level and aggregate learning}
\usage{
guesstimate(transmatrix = NULL)
}
\arguments{
\item{transmatrix}{transition matrix returned from \code{\link{multi_transmat}}}
}
\value{
list with two items: parameter estimates and estimates of learning
}
\description{
guesstimate
}

