% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interleave.R
\name{interleave}
\alias{interleave}
\title{Interleave}
\usage{
interleave(vec1, vec2)
}
\arguments{
\item{vec1}{first vector}

\item{vec2}{second vector}
}
\value{
interleaved vector
}
\description{
Interleave two character vectors. The output is a vector. The first entry is from the first vector.
Vectors can be of different lengths. If one is shorter than the other, entries of unmatched longer vector are left un-interleaved.
}
\examples{
t1 <- paste0("t1", letters[1:5]); t2 <- paste0("t2", letters[1:5]); interleave(t1, t2)
}

