% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi_transmat.R
\name{multi_transmat}
\alias{multi_transmat}
\title{multi_transmat: transition matrix of all the items}
\usage{
multi_transmat(item_df, subset = NULL)
}
\arguments{
\item{item_df}{interleaved dataframe containing all the items posed on pre- and post- test.}

\item{subset}{a dummy vector identifying the subset. Default is NULL.}
}
\value{
matrix with rows = total number of items + 1 (last row contains aggregate distribution across items)
number of columns = 4 when no don't know, and 9 when there is a don't know option
}
\description{
Creates a transition matrix for each item.
Needs an 'interleaved' dataframe (see interleave function). Pre-test item should be followed by corresponding post-item item etc.
Don't knows must be coded as NA. Function handles items without don't know responses.
The function is used internally. It calls transmat.
}
\examples{
pre_test <- data.frame(pre_item1=c(1,0,0,1,0), pre_item2=c(1,NA,0,1,0));
pst_test <- data.frame(pst_item1=pre_test[,1] + c(0,1,1,0,0), pst_item2 = pre_test[,2] + c(0,1,0,0,1))
test <- data.frame(pre_test, pst_test);
pre <- names(pre_test); pst <- names(pst_test); item_df <- test[,interleave(pre, pst)]
multi_transmat(item_df)
}

