% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{new_guess_cv}
\alias{new_guess_cv}
\title{Create a guess_cv object}
\usage{
new_guess_cv(
  fold_results,
  mean_ll,
  total_ll,
  perplexity,
  se,
  cv_type,
  k,
  call = NULL
)
}
\arguments{
\item{fold_results}{data.frame of per-fold results}

\item{mean_ll}{mean log-likelihood}

\item{total_ll}{total log-likelihood}

\item{perplexity}{perplexity}

\item{se}{standard error of perplexity across folds}

\item{cv_type}{"items" or "individuals"}

\item{k}{number of folds}

\item{call}{original function call}
}
\value{
object of class "guess_cv"
}
\description{
Create a guess_cv object
}
\keyword{internal}
