% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nona.R
\name{nona}
\alias{nona}
\title{No NAs}
\usage{
nona(vec = NULL)
}
\arguments{
\item{vec}{Required. Numeric vector.}
}
\value{
Numeric vector
}
\description{
Converts NAs to 0s
}
\examples{
x <- c(NA, 1, 2); nona(x)
}

