% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{params}
\alias{params}
\title{Parameters of Simulated Responses to Knowledge Questions Without Don't Know}
\format{A data frame with 500 rows and 4000 variables:}
\usage{
params
}
\description{
Paramaters of the simulated Pre- and Post-Process Responses
to Knowledge Questions without Don't Know
}
\keyword{datasets}

