% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-criticism.R
\name{perplexity_individuals}
\alias{perplexity_individuals}
\title{Calculate perplexity from individual-level data}
\usage{
perplexity_individuals(lca_result, pre_test, pst_test, per_individual = FALSE)
}
\arguments{
\item{lca_result}{output from lca_cor() or lca_fit()}

\item{pre_test}{data.frame of pre-test responses}

\item{pst_test}{data.frame of post-test responses}

\item{per_individual}{logical; return per-individual perplexity?}
}
\value{
numeric scalar or vector
}
\description{
Calculate perplexity from individual-level data
}
